src/requirements.txt
pyephem
requests
pyttsx3==2.90
